pyproject.toml
[tool.poetry]
name = "pnp"
version = "0.26.1"
description = "Pull 'n' Push"
authors = ["Dennis Muth <d.muth@gmx.net>"]
maintainers = ["Dennis Muth <d.muth@gmx.net>"]
license = "MIT"
readme = "README.md"
homepage = "https://pnp.readthedocs.io/"
repository = "https://github.com/HazardDede/pnp"
documentation = "https://pnp.readthedocs.io/"
keywords = ["push", "pull", "data", "pipeline"]
classifiers = [
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Developers',
    'Topic :: Home Automation',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8'
]
packages = [
    { include = "pnp" }
]
include = [
    "CHANGELOG.md",
    "LICENSE"
]

[tool.poetry.scripts]
pnp = 'pnp.console.pnp:main'
pnp_gmail_tokens = 'pnp.console.pnp_gmail_tokens:main'
pnp_record_sound = 'pnp.console.pnp_record_sound:main'


[tool.poetry.dependencies]
python = "^3.6"

# Optional "extra" ones
Adafruit_DHT = { version = "^1.3.2", optional = true }
dropbox = { version = "^9.0.0", optional = true }
face-recognition = { version = "^1.2.2", optional = true }
fitbit = { version = "^0.3.0", optional = true }
fritzconnection = { version = "^1.2.0", optional = true, python = "^3.6" }
image = { version = "^1.5.24", optional = true }
miflora = { version = "^0.4.0", optional = true }
numpy = { version = "^1.16.0", optional = true }
PyAudio = { version = "^0.2.11", optional = true }
pyftpdlib = { version = "^1.5.0", optional = true }
"RPi.GPIO" = { version = "^0.6.5", optional = true }
scipy = { version = "^1.2.0", optional = true }
speedtest-cli = { version = "^2.1.0", optional = true }
urllib3 = { version = "^1.20", optional = true }
watchdog = { version = "^0.8.3", optional = true }

# Mandatory ones
asyncio = "^3.4.3"
asyncws = "^0.1"
binaryornot = "^0.4.4"
cachetools = "^4.1.0"
cronex = "^0.1.3"
dictmentor = "^0.2.2"
glom = "^19.10.0"
Jinja2 = "^2.11.2"
psutil = "^5.7.0"
python-box = "<=3.4.6"
pytz = "^2020.1"
influxdb = "^5.3.0"
paho-mqtt = "^1.5.0"
pathspec = "^0.8.0"
requests = "^2.23.0"
"ruamel.yaml" = "^0.16.10"
schedule = "^0.6.0"
schema = "^0.7.2"
slacker = "^0.14.0"
syncasync = "^20180812"
typing-extensions = "^3.7.4"
tzlocal = "^2.0.0"
aiohttp = "^3.6.2"
typeguard = "^2.7.1"
async_generator = "^1.10"
click = "^7.1.2"
sty = "^1.0.0-beta.12"
coloredlogs = "^14.0"
fastapi = "^0.61.2"
uvicorn = "^0.12.2"
starlette_exporter = "^0.6.0"
pytest-mock = "^3.3.1"
fastcore = "^1.3.13"

[tool.poetry.dev-dependencies]
astroid = "^2.4.0"
bumpversion = "^0.5.3"
coveralls = "^2.0.0"
flake8 = "^3.7.9"
mock = "<=3.0.5"
mypy = "^0.770"
pylint = "^2.5.0"
pytest = "^5.4.1"
pytest-asyncio = "^0.11.0"
pytest-cov = "^2.8.1"
yamllint = "^1.23.0"
websockets = ">=6.0,<7.0"
invoke = "^1.4.1"
argresolver = "^0.3.3"
sphinx = "^3.1.1"
httpx = "^0.16.1"

[tool.poetry.extras]
dht = ['Adafruit_DHT']
dropbox = ['dropbox', 'urllib3']
faceR = ['image', 'face-recognition']
fitbit = ['fitbit']
fritz = ['fritzconnection']
fswatcher = ['watchdog']
ftp = ['pyftpdlib']
gpio = ['RPi.GPIO']
miflora = ['miflora']
sound = ['PyAudio', 'numpy', 'scipy']
speedtest = ['speedtest-cli']


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
